httpRequestSenderAndReceiverTest/requirements.txt
flask==2.3.2
requests==2.31.0
pillow==9.5.0
numpy==1.24.3
